Adding Playlist Support.
The TrainingRecordResource now builds a playlist of videos for each
user. The playlist is a map of video id to meta-data (completion date,
completed, required). Some videos are required, others are not.

TrainingPageResource now redirects to the earliest uncompleted chapter,
based on information in the playlist.

SurveyResultsResource now caches computed scores.