Fixing tomcat config overrides.
Finally fixing config overrides. Found a bug? in restlet where the
Servlet context parameters aren't passed to the final application. I'm
working around the issue by creating my own component which builds the
config and shares it with the frontend and backend. Overall I think it's
a much more elegant solution.

Now we have the ability to specify an external config file for the F1
credentials and the ability configure different domains in different
tomcat containers.