Changing the scoring process for slider questions.
Previously sliders were scored as a value between 1 and 4 based on the
percentage across the screen that the slider was slid. This required you
to slide fully to the right to get a 4.

Now the slider is divided into n sections, where n is the number of answers
on the slider. With 4 answers the score will be exactly 1, 2, 3, or 4.
With more answers you may get a decimals, but the last section of the
slider will always result in a 4.