Attempt to update chapter completion on every request.
I'm still seeing occasional users with missing progress information in
CCB. Currently, the progress is only updated when the user completes a
chapter. With this change, the progress will be updated every time the
training page is loaded.

CCBProgressReporter does not actually update the user's profile unless
the new level is greater than the existing level. This provision will
prevent needlessly calling the CCB API.

Additionally, this update will execute in a background thread to avoid
impacting the page load.