Add rename-videos.py script
The purpose of this script is to update the video metadata for some videos.

Usage:
   scripts/rename-videos.py <csv-file>

This script will read the csv file given on the command line.
For each row, it excepts to find:
  * video id (e.g. seeker-1),
  * new filename (without extension),
  * new duration.
The script will find each video id in the videos/ directory and rename all
the things.